task 0

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -7,7 +7,7 @@
 hash_table_t *hash_table_create(unsigned long int size)
 {
 hash_table_t *table;
-unsigned int x;
+unsigned long int x;
 if (size == 0)
 return (NULL);
 table = malloc(sizeof(hash_table_t));
